refactor(monitor): extract MonitorConfigModal + live/time composables (#438)
To hit the <300-line target on RingBufferView.vue and unlock unit testing
of the live-queue + time-filter-payload logic in isolation, split three
chunks out of the view:

* gui/src/views/ringbuffer/MonitorConfigModal.vue — the entire Monitor-
  konfigurieren modal, its form state, /stats fetch on open, and the
  POST /ringbuffer/config submit handler. RingBufferView now mounts it
  as ``<MonitorConfigModal v-model="showConfig" />`` and is unaware of
  configForm / size-unit helpers / retention-unit helpers.
* gui/src/composables/useLiveQueue.js — pause/resume + batched flush of
  WebSocket pushes onto an entries ref. Exposes paused, queuedCount,
  enqueue, pause, resume, clear, dispose.
* gui/src/composables/useTimeFilterPayload.js — pure timeFilterToPayload
  helper that converts TimeFilterPopover state (#432) into the backend
  time-filter shape (ISO bounds + signed relative seconds; point ± span
  collapses to from/to).

RingBufferView.vue: 1346 → 297 lines.
