+++ gui/src/views/RingBufferView.vue
@@ -8,7 +8,7 @@
           Speicherverhalten: file-only mit festen serverseitigen Limits.
         </p>
       </div>
-      <button @click="openConfigModal" class="btn-secondary btn-sm" data-testid="btn-open-monitor-config">⚙ Konfigurieren</button>
+      <button @click="showConfig = true" class="btn-secondary btn-sm" data-testid="btn-open-monitor-config">⚙ Konfigurieren</button>
       <button @click="applyFilters" class="btn-secondary btn-sm" data-testid="btn-refresh-ringbuffer">↻ Aktualisieren</button>
       <button
         v-if="!paused"
@@ -106,136 +106,27 @@
       </div>
     </div>
 
-    <Modal v-model="showConfig" title="Monitor konfigurieren" max-width="md">
-      <form @submit.prevent="saveConfig" class="flex flex-col gap-4">
-        <div class="grid grid-cols-1 md:grid-cols-2 gap-3">
-          <div class="flex flex-col gap-1">
-            <label class="text-xs text-slate-500">Speichermodell</label>
-            <input class="input" :value="configForm.storage" disabled />
-          </div>
-          <div class="flex flex-col gap-1">
-            <label class="text-xs text-slate-500">Max. Einträge</label>
-            <input v-model.number="configForm.max_entries" type="number" min="100" max="1000000" step="100" class="input" data-testid="rb-config-max-entries" />
-          </div>
-        </div>
-
-        <div class="rounded-lg border border-slate-200 dark:border-slate-700 p-3 flex flex-col gap-3">
-          <div class="flex items-center gap-2">
-            <input id="max-size-enabled" type="checkbox" v-model="configForm.maxSizeEnabled" />
-            <label for="max-size-enabled" class="text-sm font-medium">Max. Speicherplatz auf Platte</label>
-          </div>
-          <div class="grid grid-cols-2 gap-2">
-            <input
-              v-model.trim="configForm.maxSizeValue"
-              type="number"
-              min="1"
-              step="1"
-              class="input"
-              :disabled="!configForm.maxSizeEnabled"
-              data-testid="rb-config-max-size-value"
-              placeholder="z. B. 500"
-            />
-            <select
-              v-model="configForm.maxSizeUnit"
-              class="input"
-              :disabled="!configForm.maxSizeEnabled"
-              data-testid="rb-config-max-size-unit"
-            >
-              <option value="mb">MB</option>
-              <option value="gb">GB</option>
-            </select>
-          </div>
-        </div>
-
-        <div class="rounded-lg border border-slate-200 dark:border-slate-700 p-3 flex flex-col gap-3">
-          <div class="flex items-center gap-2">
-            <input id="retention-enabled" type="checkbox" v-model="configForm.retentionEnabled" />
-            <label for="retention-enabled" class="text-sm font-medium">Max. Retention</label>
-          </div>
-          <div class="grid grid-cols-2 gap-2">
-            <input
-              v-model.trim="configForm.retentionValue"
-              type="number"
-              min="0"
-              step="1"
-              class="input"
-              :disabled="!configForm.retentionEnabled"
-              data-testid="rb-config-retention-value"
-              placeholder="z. B. 30"
-            />
-            <select
-              v-model="configForm.retentionUnit"
-              class="input"
-              :disabled="!configForm.retentionEnabled"
-              data-testid="rb-config-retention-unit"
-            >
-              <option value="days">Tage</option>
-              <option value="months">Monate</option>
-              <option value="years">Jahre</option>
-            </select>
-          </div>
-        </div>
-
-        <div class="rounded-lg border border-slate-200 dark:border-slate-700 p-3 flex flex-col gap-2">
-          <h4 class="text-sm font-semibold">Ringbuffer Statistik</h4>
-          <div class="text-xs text-slate-500 flex items-center justify-between">
-            <span>Einträge</span>
-            <span class="font-medium text-slate-700 dark:text-slate-200" data-testid="rb-config-stats-total">{{ configStats?.total ?? '-' }}</span>
-          </div>
-          <div class="text-xs text-slate-500 flex items-center justify-between">
-            <span>Belegter Speicherplatz</span>
-            <span class="font-medium text-slate-700 dark:text-slate-200" data-testid="rb-config-stats-file-size">{{ formatBytes(configStats?.file_size_bytes ?? 0) }}</span>
-          </div>
-          <div class="text-xs text-slate-500 flex items-center justify-between">
-            <span>Effektive Retention</span>
-            <span class="font-medium text-slate-700 dark:text-slate-200" data-testid="rb-config-stats-retention">{{ formatRetention(configStats?.max_age ?? null) }}</span>
-          </div>
-        </div>
-
-        <div v-if="configMsg" :class="['p-3 rounded-lg text-sm', configMsg.ok ? 'bg-green-500/10 text-green-400' : 'bg-red-500/10 text-red-400']">
-          {{ configMsg.text }}
-        </div>
-        <div class="flex justify-end gap-3">
-          <button type="button" @click="showConfig = false" class="btn-secondary">Schliessen</button>
-          <button type="submit" class="btn-primary" :disabled="configSaving" data-testid="rb-config-save">
-            <Spinner v-if="configSaving" size="sm" color="white" />
-            Speichern
-          </button>
-        </div>
-      </form>
-    </Modal>
-
+    <MonitorConfigModal v-model="showConfig" />
   </div>
 </template>
 
 <script setup>
-import { ref, reactive, computed, onMounted, onUnmounted, nextTick } from 'vue'
+import { ref, computed, onMounted, onUnmounted, nextTick } from 'vue'
 import { ringbufferApi } from '@/api/client'
 import { useTz } from '@/composables/useTz'
 import { useSetColors } from '@/composables/useSetColors'
+import { useLiveQueue } from '@/composables/useLiveQueue'
+import { timeFilterToPayload } from '@/composables/useTimeFilterPayload'
 import { useWebSocketStore } from '@/stores/websocket'
 import Badge from '@/components/ui/Badge.vue'
 import Spinner from '@/components/ui/Spinner.vue'
-import Modal from '@/components/ui/Modal.vue'
 import TimeFilterPopover from '@/components/ui/TimeFilterPopover.vue'
 import TopbarFilterChips from '@/views/ringbuffer/TopbarFilterChips.vue'
 import TopbarStats from '@/views/ringbuffer/TopbarStats.vue'
 import FilterEditor from '@/views/ringbuffer/FilterEditor.vue'
+import MonitorConfigModal from '@/views/ringbuffer/MonitorConfigModal.vue'
 
-const LIVE_BATCH_SIZE = 200
-const LIVE_FLUSH_INTERVAL_MS = 60
-const LIVE_QUEUE_MAX = 5000
 const DEFAULT_QUERY_LIMIT = 500
-
-const SIZE_UNIT_FACTORS = {
-  mb: 1024 * 1024,
-  gb: 1024 * 1024 * 1024,
-}
-const RETENTION_UNIT_SECONDS = {
-  days: 24 * 60 * 60,
-  months: 30 * 24 * 60 * 60,
-  years: 365 * 24 * 60 * 60,
-}
 
 const { fmtDateTime } = useTz()
 const wsStore = useWebSocketStore()
@@ -252,7 +143,6 @@ function rowMatchTitle(matchedIds) {
 }
 
 const entries = ref([])
-const configStats = ref(null)
 const loading = ref(false)
 const listError = ref('')
 const showConfig = ref(false)
@@ -285,9 +175,7 @@ async function onTopbarChanged() {
   await load()
 }
 
-// TimeFilterPopover state (#432 / #438). null = no time filter active.
-// Shape: { mode: 'range', from?: Date|{seconds,sign}, to?: Date|{seconds,sign} }
-//      | { mode: 'point', point: Date|{seconds,sign}, span: {seconds,sign} }
+// TimeFilterPopover state (#432). See useTimeFilterPayload for the shape.
 const timeFilter = ref(null)
 
 async function onTimeFilterChanged() {
@@ -295,135 +183,21 @@ async function onTimeFilterChanged() {
   await load()
 }
 
-const configSaving = ref(false)
-const configMsg = ref(null)
 const tableWrapRef = ref(null)
-const paused = ref(false)
-const liveQueue = ref([])
-
 const filtersets = ref([])
 
-const configForm = reactive({
-  storage: 'file',
-  max_entries: 10000,
-  maxSizeEnabled: false,
-  maxSizeValue: '500',
-  maxSizeUnit: 'mb',
-  retentionEnabled: false,
-  retentionValue: '30',
-  retentionUnit: 'days',
-})
+const { paused, queuedCount, enqueue: enqueueLive, pause: pauseLive, resume: resumeLive, clear: clearLiveQueue, dispose: disposeLiveQueue } =
+  useLiveQueue(entries, {
+    maxEntries: DEFAULT_QUERY_LIMIT,
+    onFlush: async () => {
+      await nextTick()
+      if (tableWrapRef.value) tableWrapRef.value.scrollTop = 0
+    },
+  })
 
 const wsConnected = computed(() => wsStore.connected)
-const queuedCount = computed(() => liveQueue.value.length)
 
 let unregisterRb = null
-let liveFlushTimer = null
-
-function formatBytes(rawBytes) {
-  const bytes = Number(rawBytes)
-  if (!Number.isFinite(bytes) || bytes <= 0) return '0 B'
-  if (bytes >= SIZE_UNIT_FACTORS.gb) return `${(bytes / SIZE_UNIT_FACTORS.gb).toFixed(2)} GB`
-  if (bytes >= SIZE_UNIT_FACTORS.mb) return `${(bytes / SIZE_UNIT_FACTORS.mb).toFixed(2)} MB`
-  if (bytes >= 1024) return `${(bytes / 1024).toFixed(2)} KB`
-  return `${Math.round(bytes)} B`
-}
-
-function formatRetention(rawSeconds) {
-  const seconds = Number(rawSeconds)
-  if (!Number.isFinite(seconds) || seconds <= 0) return 'unbegrenzt'
-  if (seconds % RETENTION_UNIT_SECONDS.years === 0) return `${seconds / RETENTION_UNIT_SECONDS.years} Jahre`
-  if (seconds % RETENTION_UNIT_SECONDS.months === 0) return `${seconds / RETENTION_UNIT_SECONDS.months} Monate`
-  if (seconds % RETENTION_UNIT_SECONDS.days === 0) return `${seconds / RETENTION_UNIT_SECONDS.days} Tage`
-  return `${seconds} Sekunden`
-}
-
-function parseNonNegativeInteger(raw) {
-  const parsed = Number.parseInt(String(raw ?? '').trim(), 10)
-  if (!Number.isFinite(parsed) || parsed < 0) return null
-  return parsed
-}
-
-function pickSizeUnit(bytes) {
-  if (bytes % SIZE_UNIT_FACTORS.gb === 0) {
-    return { value: String(bytes / SIZE_UNIT_FACTORS.gb), unit: 'gb' }
-  }
-  return { value: String(Math.max(1, Math.round(bytes / SIZE_UNIT_FACTORS.mb))), unit: 'mb' }
-}
-
-function pickRetentionUnit(seconds) {
-  if (seconds % RETENTION_UNIT_SECONDS.years === 0) {
-    return { value: String(seconds / RETENTION_UNIT_SECONDS.years), unit: 'years' }
-  }
-  if (seconds % RETENTION_UNIT_SECONDS.months === 0) {
-    return { value: String(seconds / RETENTION_UNIT_SECONDS.months), unit: 'months' }
-  }
-  if (seconds % RETENTION_UNIT_SECONDS.days === 0) {
-    return { value: String(seconds / RETENTION_UNIT_SECONDS.days), unit: 'days' }
-  }
-  return { value: String(Math.ceil(seconds / RETENTION_UNIT_SECONDS.days)), unit: 'days' }
-}
-
-function hydrateConfigFormFromStats(currentStats) {
-  configForm.storage = 'file'
-  configForm.max_entries = Number(currentStats?.max_entries ?? 10000)
-
-  const maxFileSize = Number(currentStats?.max_file_size_bytes)
-  if (Number.isFinite(maxFileSize) && maxFileSize > 0) {
-    const picked = pickSizeUnit(maxFileSize)
-    configForm.maxSizeEnabled = true
-    configForm.maxSizeValue = picked.value
-    configForm.maxSizeUnit = picked.unit
-  } else {
-    configForm.maxSizeEnabled = false
-    configForm.maxSizeValue = '500'
-    configForm.maxSizeUnit = 'mb'
-  }
-
-  const maxAge = Number(currentStats?.max_age)
-  if (Number.isFinite(maxAge) && maxAge > 0) {
-    const picked = pickRetentionUnit(maxAge)
-    configForm.retentionEnabled = true
-    configForm.retentionValue = picked.value
-    configForm.retentionUnit = picked.unit
-  } else {
-    configForm.retentionEnabled = false
-    configForm.retentionValue = '30'
-    configForm.retentionUnit = 'days'
-  }
-}
-
-function buildConfigPayload() {
-  const maxEntries = Number(configForm.max_entries)
-  if (!Number.isFinite(maxEntries) || maxEntries < 100) {
-    throw new Error('Max. Einträge muss mindestens 100 sein')
-  }
-
-  const payload = {
-    storage: 'file',
-    max_entries: Math.round(maxEntries),
-    max_file_size_bytes: null,
-    max_age: null,
-  }
-
-  if (configForm.maxSizeEnabled) {
-    const sizeValue = parseNonNegativeInteger(configForm.maxSizeValue)
-    if (sizeValue === null || sizeValue <= 0) {
-      throw new Error('Max. Speicherplatz muss grösser als 0 sein')
-    }
-    payload.max_file_size_bytes = sizeValue * SIZE_UNIT_FACTORS[configForm.maxSizeUnit]
-  }
-
-  if (configForm.retentionEnabled) {
-    const retentionValue = parseNonNegativeInteger(configForm.retentionValue)
-    if (retentionValue === null) {
-      throw new Error('Retention muss eine gültige Zahl sein')
-    }
-    payload.max_age = retentionValue * RETENTION_UNIT_SECONDS[configForm.retentionUnit]
-  }
-
-  return payload
-}
 
 function extractErrorMessage(error, fallback) {
   return error?.response?.data?.detail || error?.message || fallback
@@ -441,93 +215,15 @@ async function loadFiltersets() {
   }
 }
 
-function enqueueLive(entry) {
-  liveQueue.value.push(entry)
-  if (liveQueue.value.length > LIVE_QUEUE_MAX) {
-    liveQueue.value.splice(0, liveQueue.value.length - LIVE_QUEUE_MAX)
-  }
-  if (!paused.value) scheduleLiveFlush()
-}
-
-function scheduleLiveFlush() {
-  if (paused.value || liveFlushTimer) return
-  liveFlushTimer = setTimeout(() => {
-    liveFlushTimer = null
-    void flushLiveQueue()
-  }, LIVE_FLUSH_INTERVAL_MS)
-}
-
-async function flushLiveQueue() {
-  if (paused.value || !liveQueue.value.length) return
-  const batch = liveQueue.value.splice(0, LIVE_BATCH_SIZE)
-  if (batch.length) {
-    entries.value = [...batch.reverse(), ...entries.value].slice(0, DEFAULT_QUERY_LIMIT)
-    await nextTick()
-    if (tableWrapRef.value) tableWrapRef.value.scrollTop = 0
-  }
-  if (liveQueue.value.length) scheduleLiveFlush()
-}
-
-function buildTimeFilter() {
-  // Convert the TimeFilterPopover state (#432) into the backend time-filter
-  // shape:
-  //   { from?: iso, to?: iso,
-  //     from_relative_seconds?: int, to_relative_seconds?: int }
-  // Date bounds → ISO strings; relative durations → signed seconds.
-  // Point mode (point ± span) collapses into an absolute (from, to) pair.
-  const filter = timeFilter.value
-  if (!filter) return null
-  const time = {}
-
-  function applyBound(bound, key, relKey) {
-    if (!bound) return
-    if (bound instanceof Date) {
-      time[key] = bound.toISOString()
-    } else if (Number.isFinite(bound.seconds)) {
-      time[relKey] = (bound.sign === -1 ? -1 : 1) * bound.seconds
-    }
-  }
-
-  if (filter.mode === 'point') {
-    const point = filter.point instanceof Date
-      ? filter.point
-      : (Number.isFinite(filter.point?.seconds)
-          ? new Date(Date.now() + (filter.point.sign === -1 ? -1 : 1) * filter.point.seconds * 1000)
-          : null)
-    const spanSeconds = Number.isFinite(filter.span?.seconds) ? filter.span.seconds : 0
-    if (point && spanSeconds > 0) {
-      time.from = new Date(point.getTime() - spanSeconds * 1000).toISOString()
-      time.to = new Date(point.getTime() + spanSeconds * 1000).toISOString()
-    } else if (point) {
-      time.from = point.toISOString()
-      time.to = point.toISOString()
-    }
-  } else {
-    applyBound(filter.from, 'from', 'from_relative_seconds')
-    applyBound(filter.to, 'to', 'to_relative_seconds')
-  }
-
-  return Object.keys(time).length ? time : null
-}
-
 function buildQueryV2() {
   const payload = {
     filters: {},
     sort: { field: 'ts', order: 'desc' },
     pagination: { limit: DEFAULT_QUERY_LIMIT, offset: 0 },
   }
-  const time = buildTimeFilter()
+  const time = timeFilterToPayload(timeFilter.value)
   if (time) payload.filters.time = time
   return payload
-}
-
-function pauseLive() {
-  paused.value = true
-}
-
-function resumeLive() {
-  paused.value = false
-  scheduleLiveFlush()
 }
 
 function onLiveEntry(entry) {
@@ -538,15 +234,14 @@ function onLiveEntry(entry) {
 }
 
 async function applyFilters() {
-  liveQueue.value = []
+  clearLiveQueue()
   await load()
 }
 
 onMounted(async () => {
   // Load filtersets first so the topbar-colour cache is populated before
-  // load() decides between queryV2 and queryMultiFiltersets (#437).
-  // /stats is no longer fetched eagerly — TopbarStats owns its own fetch
-  // and the config modal pulls fresh stats on demand via openConfigModal().
+  // load() picks between queryV2 and queryMultiFiltersets (#437). /stats
+  // is owned by TopbarStats / the config modal — not fetched here.
   await loadFiltersets()
   await load()
   unregisterRb = wsStore.onRingbufferEntry(onLiveEntry)
@@ -554,7 +249,7 @@ onMounted(async () => {
 
 onUnmounted(() => {
   unregisterRb?.()
-  clearTimeout(liveFlushTimer)
+  disposeLiveQueue()
 })
 
 function activeTopbarSetIds() {
@@ -576,7 +271,7 @@ async function load() {
       // Multi-filterset OR-union query (#431); each entry carries
       // matched_set_ids so the table can colour-code rows by matched set.
       const body = { set_ids: setIds }
-      const time = buildTimeFilter()
+      const time = timeFilterToPayload(timeFilter.value)
       if (time) body.time = time
       const resp = await ringbufferApi.queryMultiFiltersets(body)
       data = resp.data
@@ -593,41 +288,6 @@ async function load() {
     listError.value = extractErrorMessage(error, 'Monitor-Abfrage fehlgeschlagen')
   } finally {
     loading.value = false
-  }
-}
-
-async function loadStatsForConfig() {
-  // On-demand /stats fetch invoked by openConfigModal — the config form
-  // needs current limits/retention plus the "Belegter Speicherplatz" /
-  // "Effektive Retention" read-outs in the modal.
-  try {
-    const { data } = await ringbufferApi.stats()
-    configStats.value = data
-    hydrateConfigFormFromStats(data)
-  } catch {
-    // Silent on failure; the modal still renders with the configForm
-    // defaults and the user can save without read-back values.
-  }
-}
-
-async function openConfigModal() {
-  showConfig.value = true
-  await loadStatsForConfig()
-}
-
-async function saveConfig() {
-  configSaving.value = true
-  configMsg.value = null
-  try {
-    const payload = buildConfigPayload()
-    const { data } = await ringbufferApi.config(payload)
-    configStats.value = data
-    hydrateConfigFormFromStats(data)
-    configMsg.value = { ok: true, text: 'Monitor-Konfiguration gespeichert' }
-  } catch (error) {
-    configMsg.value = { ok: false, text: extractErrorMessage(error, error?.message || 'Speichern fehlgeschlagen') }
-  } finally {
-    configSaving.value = false
   }
 }
